Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <math.h>
 int main(){
-    int a,b , c, M;
-    scanf("%d%d%d",&a,&b,&c);
+    float a,b , c, M;
+    scanf("%f%f%f",&a,&b,&c);
     M =( a +b+c)/3;
-    printf("Average: %d",M);
+    printf("Average: %f",M);
     return 0;
 }
